Fix override message showing upon invalid processing
**CONTEXT:**
When a visit with a rejection reason automatically
checked by NOMIS and staff would uncheck the rejection (effectively overriding
the rejection reason) but did not either select a date or another rejection. The
processing would fail with a validation message. However the override
restriction warning would be shown upon redisplaying the page which was
incorrect.

**CHANGES:**

The main challenge was to pass down the validation state of the StaffResponse
instance.
I had to extract the instanciation of the StaffResonse object from within the
BookingResponder and inject it at the StaffResponseContext level.