Fix autoloading bug in dev environment
Following the removal of the unnecessary types (PR:
https://github.com/ministryofjustice/prison-visits-2/pull/1623) a bug
was introduced into the development environment. If a change was made to
the codebase and a page refreshed it would show the following error:
"A copy of ContactsEnumerableType has been removed from the module tree
but is still active!"

Investigation found that this error was related to the types being in
the autoload path and therefore the 'types' folder has now been moved to
the 'lib' folder instead.