Update cookie expiry
Up until this point there has been no expiry time for the staff (pvb2)
cookie.  Following advice from the security team we have now set the
cookie to expire after an hour.  The way that Rails manages the
session_store means that active use of the site will keep extending the
expiry time and will therefore only timeout once a user closes the
browser, or doesn't use the site for an hour.