Restrict print visits search to 6 months (#1715)
At present staff are able to enter a date into the print visits and it
will return any visits as far back as they want to search.  However, we
want to establish whether or not staff actually need to search for
visits from further back than 6 months.  As such we are going to restrict
their searches to the last 6 months and advise them to contact us if
they need more historical visits data.