Fix print visits feature spec
We recently implemented a feature that only allows staff to search for,
and print visits within the last 6 months.  As such one of the feature
tests is now failing as the date falls outside of this range.  The test
has been updated to be more dynamic and should prevent this from
happening again in the future.