Fix print visits feature spec
We recently implemented a feature that only allows staff to print visits
either in the future, or up until six months in the past.  As a result
one of the print visits feature specs is failing as the date now falls
outside of this range.  The test is now more dynamic and should prevent
this from happening again in the future.