Remove Book To Nomis code
Following a decision to no longer continue with the development and
roll out of Book To Nomis we are going to remove the code to reduce
cofusion and complexity.

The first step was to remove the configuration and feature flags related
to Book To Nomis and get all tests passing again.