Fix print visits feature spec
In the fix_print_visits_feature_spec we fill in certain fields with
various dates to test that the right messages are shown to staff.  We
recently implemented a feature that only lets staff search/print visits
from the last 6 months.  As a result a test has started failing as now
falls outside of this date.  We are now managing this date dynamically to
avoid this issue in the future.