Remove 'prisoner banned' checkbox (#1776)
We are in the process of getting a change made to one of the NOMIS API
endpoints as it is incorrectly preventing visitors from being able to
request online visits if the prisoner is banned.  In anticipation of this
we are also removing the prisoner banned checkbox on the staff site as
this should also not be an option to reject a visit.