Fix stats for rejection reasons (#1781)
Following the removal of the 'prisoner banned' rejection reason for staff
there has been a knock on impact for the stats page as the rejection
percentages rely on the rejection reasons; without 'prisoner_banned'
being in that list it fails with the error message, "undefined method
prisoner_banned=".  This is because we're creating methods based on that
list, and as we need to display historical data we cannot get rid of the
reason entirely.