Remove restrictions feature
As part of a change in the PVB team's focus we are removing features
that we have not been able to fully test or rollout across the estate.

We were aiming to display restrictions for both prisoner and their
visitor(s), but are now removing this code to avoid complexity and confusion.