REMOVE Prisoner Location validation for PVB Public
PVB2 is making a call to the Nomis API offender location  endpoint on
behalf of PVB Public endpoint when it does not need to. In addition, the
Nomis offender location end point is currently problematic (ticket has been raised for
possible database contention) and resulting in time outs.

Prisoner location validation is therefore removed from the
ApiPrisonerChecker class.