Restrict access to estates when there is no match
Fixes a bug that gave access to an estate that has no match on the pvb
admin configuration.

The bug was found when it was noticed that a value on the pvb admin
configuration on the yaml had a mispelling, but users still got access
to the estate referenced by the users sso permissions.

As part of this fix more mocks have been added to the test suite to get
it green as before the test suite took advantage of the bug to give
access to estates for users with no match.