Pull Request template
As a team we have been discussing how we can be consistent around our
git styles.  One of the tasks is to create a PR template which will help
the team with this consistency, and act as a prompt for ensuring things
like the documentation has been updated, or the deploy repo has been
updated.

This is a first pass and will need to be reviewed by the wider team as
there are not only devs that create PRs, and we want this to be useful/
meaningful for everyone.

Trello card: https://trello.com/c/uBKazmeE/178-use-gds-git-style-guide