Fix did_you_mean issue on circleci
Tests are executed through rake in circle, for some unknown reason the
`did_you_mean` gem stopped being available when the specs run.

The fix requires to update the gem version that circle comes with as it
is one version behind and to include it on the Gemfile so that it gets
required for the test run.