Update Dartmoor visit slots
Dartmoor are in the process of changing some of their social visit slot
days and times and want this to come into affect 30th September.  To ensure
our calendar shows the correct slots during this changeover period we
need to add some unbookable and anomalous dates.  A second PR will be
created which will need to be merged after 30th September to fully
implement the new days/times.