Reverse changes to Swansea slots
A request has come in from Swansea to make significant changes to their
slot times.  A PR was created
(https://github.com/ministryofjustice/prison-visits-2/pull/2036),
merged and deployed to staging, but it has since been established that
these changes aren't due to come into effect immediately. Therefore we
need to reverse this commit and only merge/deploy to production when
required.  There are some other prisons that require their slots updated
immediately and therefore we cannot leave the changes to Swansea in
staging.