Remove references to amended booking window
When The Verne initially contacted us they requested that visitors be
able to book visits up to 6 months in advance, so help them book travel
and accommodation in a timely manner.  However, there are restrictions
in NOMIS and the VO which means that this can't be done.  Therefore,
this PR removes reference to increasing the booking window to avoid
confusion in the future.