Add steps to CI pipeline
We are using CircleCi for our deployment pipeline and at present we have
got to the point where we are able to deploy to staging.  However, we
need to be able to test staging, and if successful then require someone
to click a button to confirm they want to deploy to production.  This PR
adds those latter steps into the workflow.