Set deploy to prod step in pipeline
As we prepare for the migration of PVB from template deploy to
kubernetes we need to add the ability for changes to be deployed to
production.

NOTE: We are unable to currently include a 'test staging' step, please
see https://github.com/ministryofjustice/prison-visits-public/pull/748
for details behind this decision.