Add HMP The Verne
Following a visit to The Verne we temporarily removed them from our
production environment to give them the opportunity to finalise
processes at their end, and get the slots named correctly.  They have
now contacted us again and we are ready to test their set up again.

We have also established that they are not a private prison as we
originally thought, and have requested a 4 day lead in window.