Add flash alert re: service downtime
We are due to migrate PVB from template deploy to kubernetes by the end
of the week.  An email is being sent out to staff to alert them to this
news, however we are also going to add a flash alert that will be
displayed at the start of each session for that user.  This will
disappear as soon as they move on to another page, so not to be a
constant distraction.

This flash alert PR will need to be reverted as part of the migration
tasks to ensure staff do not see it once the work has been completed.