Update service files in deployment
Whilst investigating an issue I noticed that we had failed to add the
metrics port in the service file in both staging and production.  This
means that logs are not able to be sent to Kibana and we see the
following error instead:

"Prometheus Exporter, failed to send message Cannot assign requested
address - connect(2) for "localhost" port 9394".

By adding the ports in we should now start seeing all logs coming
through to Kibana.