Increase failed replica count for cronjob
At present it does not appear that our cron jobs in prod or staging are
working as expected.  For example we have them scheduled to run every 24
hours, but in our staging env the last 3 pods are from 2, 4 & 7 days
ago.  I want to increase the number of failed pods to see if this helps
bring more visibility to this issue.