Activate HMP The Verne
We have been contacted by staff at The Verne to confirm that they are
now ready to start using the Prison Visits Booking service.  They have
provided us with a booking line and advised of the opening times, which
has been added to this PR.  I have also removed all the unbookable dates
so that visitors will be able to start using the service immediately.

Once the service has gone live on Monday we just need to inform staff so
that they can monitor the site for visit requests.