Increase memory limit for K8s
We are continuing to work to find the right levels for our kubernetes
set up to ensure that the service is available/stable for our users.
Following a conversation with Cloud Platform we have made changes to our
namespace and are now increasing the memory limit in staging to 2GI to
check that this change doesn't have any negative consequences.