15-18yr old week day slots
The slots in NOMIS for the 15-18yr old week day visits are incorrect,
returning 1 hour slots rather than 3 hours; meaning that when testing on
staging all the slots were showing as fully booked.  I went back to
Feltham to offer options on how to fix this and they have confirmed that
they want us to show the 1 hour slot; they will then advise visitors
that it's actually a 3 hour visit slot.