Document the Api.rb class
This class has several methods that require a nomis_id and some that
require and offender_id (which is an integer) so rather than rely on the
parameter names, document what needs to be passed.

There are several TODOs still in this documentation because we don't
know (and it isn't documented) what the constraints are on the date
ranges we use.  Some need to be in the future (start and end) but it
would be good/useful for this to be much clearer.