Update PVB VCR cassettes
Whilst working on migrating PVB from the legacy NOMIS API to Elite2
(https://dsdmoj.atlassian.net/browse/POM-401) we realised that we needed
to ensure that the PVB specs were up to date/accurate.  This ensures
that we can be confident when running the specs on the migration branch,
and not be in a situation where we are changing our code and tests
during the same ticket.

Please note: some of the VCR cassettes have been deleted and not
re-recorded as they are related to code that was deleted e.g. the book
to NOMIS feature.