Reset API spec
Following a conversation with colleagues in Sheffield we were able to
establish that the behaviour had indeed changed when calling the
endpoint to fetch an offender's contact list; Elite2 has been ignoring
duplications, whereas the legacy API returned all entries.  This has now
been fixed and so the spec has been reverted back to expect the original
number of contacts, 27.