Update database credentials for prod environment
Our staging environment was not working due to upgrade changes
made to k8s. Staging has been fixed by pointing our credentials
to the correct location. As the production environment is due
to be recycled also this commit is to point our credentials to
the correct location so we don't have the same problems we
encounted in staging.