Fix flaky spec
Nomis::Api makes use of the ::instance method from the Singleton module

This results in a single instance to be returned each time
Nomis::Api.instance is been called.

As Nomis::Api.enabled? is been called inside its initializer, an
exception was raised which resulted in a failing spec.

It seems that the intention of this stub was to make sure
Nomis::Api.enabled? returns false in case it was called anywhere else in the application.